[SPARK-44796][BUILD][CONNECT] Remove `grpc-java` plugin related configuration from the `connect/connect-client-jvm` module
### What changes were proposed in this pull request?
This pr remove `grpc-java` plugin related configuration from the `connect/connect-client-jvm module` for SBT because the code generation related to `grpc-java` is handled by the `connect-common` module. The reason this pr did not touch the Maven `pom.xml` is because the corresponding Maven modules originally did not have that configuration.

### Why are the changes needed?
Clean up unnecessary sbt configurations.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
- Pass GitHub Actions

Closes #42477 from LuciferYang/SPARK-44796.

Authored-by: yangjie01 <yangjie01@baidu.com>
Signed-off-by: yangjie01 <yangjie01@baidu.com>