[SPARK-44653][SQL][FOLLOWUP] ResolveUnion should not combine Unions
### What changes were proposed in this pull request?

This is a followup of https://github.com/apache/spark/pull/42315 , to fix by-name Union as well. Ideally, an analyzer rule should not invoke an optimizer rule, and it's cleaner to keep the df union hack in the df code.

### Why are the changes needed?

fix the regression for by-name Union as well.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

new tests

Closes #42483 from cloud-fan/union.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Kent Yao <yao@apache.org>