[SPARK-44803][BUILD] Replace `publish` with `publishOrSkip` in SparkBuild to eliminate warnings
### What changes were proposed in this pull request?
The pr aims to replace `publishOrSkip` with `publish` in SparkBuild to eliminate warnings.

### Why are the changes needed?
After we upgraded SBT to 1.9.1+, the `publish` method has been marked as deprecated, to reduce compilation warnings.
<img width="1290" alt="image" src="https://github.com/apache/spark/assets/15246973/ac45809d-60b4-4879-92e3-f4650f4cfbc2">

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Manual testing.
We use the submodule 'tools' as an example to detect the local repo of mvn and ivy.
command:
```
  sh build/sbt
  show tools/M2r/publishLocal
  show tools/Sbt/publishLocal
```

Closes #42486 from panbingkun/SPARK-44803.

Authored-by: panbingkun <pbk1982@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>