[SPARK-44822][PYTHON][SQL] Make Python UDTFs by default non-deterministic
### What changes were proposed in this pull request?

This PR changes the default determinism of Python UDTFs from `True` to `False`.

### Why are the changes needed?

To prevent potential regressions as many Python UDTFs are often used as non-deterministic UDTFs. Users can always mark them as deterministic.

### Does this PR introduce _any_ user-facing change?

No. Python UDTF is a new feature that is not yet released.

### How was this patch tested?

Existing and new tests

Closes #42519 from allisonwang-db/spark-44822-non-det-by-default.

Authored-by: allisonwang-db <allison.wang@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>