[SPARK-44873] Support alter view with nested columns in Hive client
### What changes were proposed in this pull request?
Previously, if a view's schema contains a nested struct, alterTable using Hive client will fail. This change supports a view with a nested struct. The mechanism is to store an empty schema when we call Hive client, since we already store the actual schema in table properties. This fix is similar to https://github.com/apache/spark/pull/37364

### Why are the changes needed?
This supports using view with nested structs in Hive metastore.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Added unit test.

Closes #42532 from kylerong-db/hive_view.

Authored-by: kylerong-db <kyle.rong@databricks.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>