[SPARK-44960][UI] Unescape and consist error summary across UI pages
### What changes were proposed in this pull request?

This pull request eliminates the unnecessary use of escape for error summary cells.

Previously, all the error details and some of the error summaries, such as the Task list on the stage page, did not call escapeHtml4. The rest of the error summaries, such as Job list, Stage list, Query list, etc., will do an extra escapeHtml4 that is unnecessary.

### Why are the changes needed?

UI improvement

### Does this PR introduce _any_ user-facing change?

UI changes

### How was this patch tested?

#### new tests
#### Local test

##### Before

![image](https://github.com/apache/spark/assets/8326978/f65e0713-087f-47a0-9cc9-321898ec6e6c)

##### After

![image](https://github.com/apache/spark/assets/8326978/e07c06f5-72ab-4465-b950-3e51350a3046)

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #42674 from yaooqinn/SPARK-44960.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: Kent Yao <yao@apache.org>