[SPARK-44832][CONNECT] Make transitive dependencies work properly for Scala Client
### What changes were proposed in this pull request?
This PR cleans up the Maven build for the Spark Connect Client and Spark Connect Common. The most important change is that we move `sql-api` from a `provided` to `compile` dependency. The net effect of this is that when a user takes a dependency on the client, all of its required (transitive) dependencies are automatically added.

Please note that this does not address concerns around creating an überjar and shading. That is for a different day :)

### Why are the changes needed?
When you take a dependency on the connect scala client you need to manually add the `sql-api` module as a dependency. This is rather poor UX.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Manually running maven, checking dependency tree, ...

Closes #42518 from hvanhovell/SPARK-44832.

Authored-by: Herman van Hovell <herman@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>