[SPARK-44995][K8S] Promote `SparkKubernetesClientFactory` to `DeveloperApi`
### What changes were proposed in this pull request?

This PR aims to promote `SparkKubernetesClientFactory` as **stable** `DeveloperApi` in order to maintain it officially in a backward compatible way at Apache Spark 4.0.0.

### Why are the changes needed?

Like SPARK-35280 and SPARK-37497, `SparkKubernetesClientFactory` is also able to be used to develop new `ExternalClusterManager` for K8s environment.
- https://github.com/apache/spark/pull/32406
- https://github.com/apache/spark/pull/34751

### Does this PR introduce _any_ user-facing change?

No. Previously, it was `private[spark]`.

### How was this patch tested?

Manual review because this is only a visibility change.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #42709 from dongjoon-hyun/SPARK-44995.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>