[SPARK-45010][INFRA] Limit GHA job execution time to up to 5 hours in `build_and_test.yml`
### What changes were proposed in this pull request?

This PR aims to limit GHA job execution time to up to 5 hours in `build_and_test.yml` in order to avoid idle hung time.
New limit is applied for all jobs except three jobs (`precondition`, `infra-image`, and `breaking-changes-buf`) which didn't get a hung situation before.

### Why are the changes needed?

Currently, the default value (6 hour) is used.

- https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idtimeout-minutes
  > The maximum number of minutes to let a job run before GitHub automatically cancels it. Default: 360

However, when the jobs hit the limit, it's already a hung job and the pipeline becomes idle for over 5 hours like the following.

- https://github.com/apache/spark/actions/runs/5999379610/job/16269462910 (IDLE for about 5 hours 3 minutes)
```
2023-08-28T12:44:10.9217219Z Starting test(python3.9): pyspark.mllib.tests.test_algorithms (temp output: /__w/spark/spark/python/target/1a1ff656-2923-4d34-b046-1e947ab0f05b/python3.9__pyspark.mllib.tests.test_algorithms__f9df3bwd.log)
2023-08-28T17:47:21.5922036Z ##[error]The operation was canceled.
```

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

No because the successful CIs finished in 4 ~ 5 hours.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #42727 from dongjoon-hyun/SPARK-45010.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>