[SPARK-45061][SS][CONNECT] Clean up Running python StreamingQueryLIstener processes when session expires
### What changes were proposed in this pull request?

Clean up all running python StreamingQueryLIstener processes when session expires

### Why are the changes needed?

Improvement

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Test will be added in SPARK-44462. Currently there is no way to test this because the session will never expire. This is because the started python listener process (on the server) will establish a connection with the server process with the same session id and ping it all the time.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #42687 from WweiL/SPARK-44433-followup-listener-cleanup.

Authored-by: Wei Liu <wei.liu@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>