[SPARK-45031][INFRA] Choose the right merge code path and merge hash for reopened PRs
### What changes were proposed in this pull request?

When determining to cherry-pick a PR, we also check the PR is in the closed state; Otherwise, we assume it gets reverted and reopened, and we go normal merging.

When cherry-picking, we shall select the merge hash from the latest commit-close event instead of the oldest. This is a bug fix

### Why are the changes needed?

Bugfix

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

mutually

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #42749 from yaooqinn/SPARK-45031.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>