[SPARK-44845][YARN][DEPLOY] Fix file system uri comparison function
### What changes were proposed in this pull request?
What changes were proposed in this pull request?

### Why are the changes needed?

In the org.apache.spark.deploy.yarn.Client#compareUri method, hdfs://hadoop81:8020 and hdfs://192.168.0.81:8020 are regarded as different file systems (hadoop81 corresponds to 192.168.0.81). The specific reason is that in the last pr, different URIs of user information are also regarded as different file systems. Uri.getauthority is used to determine the user information, but authority contains the host so the URI above must be different from authority. To determine whether the user authentication information is different, you only need to determine URI.getUserInfo.

the last pr and issue link:
https://issues.apache.org/jira/browse/SPARK-22587

https://github.com/apache/spark/pull/19885

### Does this PR introduce _any_ user-facing change?

### How was this patch tested?

Closes #42529 from zekai-li/master.

Authored-by: zekai-li <58294989+zekai-li@users.noreply.github.com>
Signed-off-by: Mridul Muralidharan <mridul<at>gmail.com>