[SPARK-44866][SQL] Add `SnowflakeDialect` to handle BOOLEAN type correctly
### What changes were proposed in this pull request?

In Snowflake a BOOLEAN data type exist but not the BIT data type.
This PR adds `SnowflakeDialect` to override the default JdbcDialect and redefine the default mapping behaviour for the _boolean_ type.  Currently, it's mapped to `BIT(1)` type.

https://github.com/apache/spark/blob/42ec07dad459a4ee04ebe2c6f4b38efb371aaab9/sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/jdbc/JdbcUtils.scala#L149

### Why are the changes needed?

The BIT type does not exist in Snowflake. This cause the Spark Job to fail on table creation.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Unit test and directly on Snowflake

Closes #42545 from hayssams/master.

Authored-by: Hayssam Saleh <Hayssam.saleh@starlake.ai>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>