[SPARK-45034][SQL] Support deterministic mode function
### What changes were proposed in this pull request?
This PR adds a new optional argument to the `mode` aggregate function to provide deterministic results. When multiple values have the same greatest frequency then the new boolean argument can be used to get the lowest or highest value instead of an arbitraty one.

### Why are the changes needed?
To make the function more user friendly.

### Does this PR introduce _any_ user-facing change?
Yes, it adds a new argument to the `mode` function.

### How was this patch tested?
Added new UTs.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #42755 from peter-toth/SPARK-45034-deterministic-mode-function.

Authored-by: Peter Toth <peter.toth@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>