[SPARK-45191][SQL] InMemoryTableScanExec simpleStringWithNodeId adds columnar info
### What changes were proposed in this pull request?

InMemoryTableScanExec supports both row-based and columnar input and output which is based on the cache serialzier. It would be more friendly for user if we can provide the columnar info to show whether it is columnar in/out.

### Why are the changes needed?

Add columnar info for InMemoryTableScanExec explain.

### Does this PR introduce _any_ user-facing change?

no, if no columnar input or output.

### How was this patch tested?

manually test a columnar supporting example.

before:
<img width="303" alt="image" src="https://github.com/apache/spark/assets/12025282/289cba7e-51af-4b01-b591-bc9c8328801d">

after:
<img width="435" alt="image" src="https://github.com/apache/spark/assets/12025282/e1514a2c-bf47-47c3-b311-23e49c9db222">

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #42967 from ulysses-you/cache.

Authored-by: ulysses-you <ulyssesyou18@gmail.com>
Signed-off-by: Xiduo You <ulyssesyou@apache.org>