[SPARK-45277][BUILD][INFRA] Install Java 17 to support SparkR testing on Windows
### What changes were proposed in this pull request?
This PR adds the step of installing Java 17 in `appveyor-install-dependencies.ps1` to make SparkR can use Java 17 for testing on Windows.

On the other hand, this pr corrects the `log4j2` configuration file name used for testing in `appveyor.yml`.

### Why are the changes needed?
Apache Spark supports Java 17 at a minimum, so SparkR also needs to use Java 17 for testing on Windows. On the other hand, the current version of Windows in use does not come with Java 17 pre-installed, so this PR carried out a manual installation.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
`continuous-integration/appveyor/pr` should test pass

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #43056 from LuciferYang/SPARK-45277.

Authored-by: yangjie01 <yangjie01@baidu.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>