[SPARK-44756][CORE] Executor hangs when RetryingBlockTransferor fails to initiate retry
### What changes were proposed in this pull request?
This PR fixes a bug in `RetryingBlockTransferor` that happens when retry initiation has failed.

With this patch, the callers of `RetryingBlockTransfeathror#initiateRetry()` will catch any error and invoke the parent listener's exception handler.

### Why are the changes needed?
This is needed to prevent an edge case where retry initiation fails and executor gets stuck.

More details in SPARK-44756

### Does this PR introduce _any_ user-facing change?
N/A

### How was this patch tested?
Added a new test case in `RetryingBlockTransferorSuite` that simulates the problematic scenario.

<img width="772" alt="image" src="https://github.com/apache/spark/assets/17327104/f20ec327-f5c9-4d74-b861-1ea4e05eb46b">

Closes #42426 from hdaikoku/SPARK-44756.

Authored-by: Harunobu Daikoku <harunobu.daikoku@agoda.com>
Signed-off-by: Mridul Muralidharan <mridul<at>gmail.com>