[SPARK-44913][SQL] DS V2 supports push down V2 UDF that has magic method
### What changes were proposed in this pull request?

Right now we only support pushing down the V2 UDF that has not a magic method. Because the V2 UDF will be analyzed into the `ApplyFunctionExpression` which could be translated and pushed down. However, a V2 UDF that has the magic method will be analyzed into `StaticInvoke` or `Invoke` that can not be translated into V2 expression and then can not be pushed down to the data source. The magic method is suggested.

### Why are the changes needed?

This PR adds the support of pushing down the V2 UDF that has a magic method.

### Does this PR introduce _any_ user-facing change?

Yes, now the V2 UDF with the magic method could be pushed down.

### How was this patch tested?

New UTs.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #42612 from ConeyLiu/push-down-udf-with-magic.

Lead-authored-by: Xianyang Liu <liu-xianyang@hotmail.com>
Co-authored-by: xianyangliu <xianyangliu@tencent.com>
Signed-off-by: Chao Sun <sunchao@apple.com>