[SPARK-45377][CORE] Handle InputStream in NettyLogger
### What changes were proposed in this pull request?

Handle `InputStream`s in the `NettyLogger` so we can print out how many available bytes there are.

### Why are the changes needed?

As part of the SSL support we are going to transfer `InputStream`s via Netty, and this functionality makes it easy to see the size of the streams in the log at a glance.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

CI. Tested as part of the changes in https://github.com/apache/spark/pull/42685 which this is split out of, I observed the logs there.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #43165 from hasnain-db/spark-tls-netty-logger.

Authored-by: Hasnain Lakhani <hasnain.lakhani@databricks.com>
Signed-off-by: Sean Owen <srowen@gmail.com>