[SPARK-45378][CORE] Add convertToNettyForSsl to ManagedBuffer
### What changes were proposed in this pull request?

As the title suggests. In addition to that API, add a config to the `TransportConf` to configure the default block size if desired.

### Why are the changes needed?

Netty's SSL support does not support zero-copy transfers. In order to support SSL using Netty we need to add another API to the `ManagedBuffer` which lets buffers return a different data type.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

CI. This will have tests added later - it's tested as part of https://github.com/apache/spark/pull/42685 from which this is split out.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #43166 from hasnain-db/spark-tls-buffers.

Authored-by: Hasnain Lakhani <hasnain.lakhani@databricks.com>
Signed-off-by: Mridul Muralidharan <mridul<at>gmail.com>