[SPARK-45404][CORE] Support `AWS_ENDPOINT_URL` env variable
### What changes were proposed in this pull request?

This PR aims to support `AWS_ENDPOINT_URL` in addition to the existing `AWS_*` environment variables.

### Why are the changes needed?

To improve the usability. This is useful when we use S3-compatible object storages.
- https://docs.aws.amazon.com/sdkref/latest/guide/settings-reference.html#EVarSettings

### Does this PR introduce _any_ user-facing change?

This is a new feature. If a user have a misconfigured `AWS_ENDPOINT_URL`, it will hang or mostly be denied during the authentication.

### How was this patch tested?

Pass the CIs with newly updated test case.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #43205 from dongjoon-hyun/SPARK-45404.

Authored-by: Dongjoon Hyun <dhyun@apple.com>
Signed-off-by: Dongjoon Hyun <dhyun@apple.com>