[SPARK-45472][SS] RocksDB State Store Doesn't Need to Recheck checkpoint path existence
### What changes were proposed in this pull request?
In RocksDBFileManager, we add a variable to indicate that root path is already checked and created if not existing, so that we don't need to recheck the second time.

### Why are the changes needed?
Right now, every time RocksDB.load() is called, we check checkpoint directory existence and create it if not. This is relatively expensive and show up in performance profiling.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Existing CI tests to cover it.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #43299 from siying/rootPath.

Authored-by: Siying Dong <siying.dong@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>