[SPARK-45204][CONNECT] Add optional ExecuteHolder to SparkConnectPlanner
### What changes were proposed in this pull request?
This PR adds an optional `ExecuteHolder` to `SparkConnectPlanner`.
This allows plugins to access the `ExecuteHolder` to for example create a `QueryPlanningTracker` without the need to create a new `CommandPlugin` interface like proposed in https://github.com/apache/spark/pull/42984.

### Why are the changes needed?
There is currently no way to track queries executed by a CommandPlugin. For this, the plugin needs access to the `ExecuteHolder`.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Adjusted existing tests.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #43311 from dillitz/plugin-new-approach.

Authored-by: Robert Dillitz <robert.dillitz@databricks.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>